requirements.txt
requests
pymongo
together
streamlit==1.24.1
openai==1.31.1
pandas==2.0.3
youtube-transcript-api==0.6.2
openpyxl==3.0.5
yt-dlp
imageio-ffmpeg
python-dotenv
deepgram-sdk==3.2.7
langchain
langchain-mongodb 
langchain-openai
langchain-community